blog/requirements.txt
Django==2.1.7
pytz==2018.9
djangorestframework==3.7.7
django-rest-auth==0.9.3
django-rest-swagger== 2.2.0
django-filter==2.0.0
factory_boy==2.11.1